fix: getCropModesテストを別ファイルに分離してCI失敗を解決
plan.test.tsのvi.resetModules()がgetCropModesのvi.mockをクリア
していたため、getCropModesテストを専用ファイルに分離。
react-image-cropのモックが独立して動作するようになりCI安定化。
